Accessibility: Hack to make AnswerIndicators accessible
Due to the inherent inaccessibility of Popper, this is a very dirty workaround to make the contents of the results table known to Screen Reader users. It shouldn't be done this way, but doing it properly would require a major architectural change. Also, aria-label can only be used on interactive elements, which this is not.